limit must be size_t
git-svn-id: https://svn.r-project.org/R/trunk@50907 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/gnuwin32/malloc.c
@@ -646,7 +646,7 @@ extern "C" {
 
 /* R Specific declarations here */
 
-    extern unsigned int R_max_memory;
+extern size_t R_max_memory;
 extern int R_Is_Running;
 static size_t R_used = 0;
 void Rf_warning(const char *, ...);
